+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## dev
 
 * add support for `django.core.cache.backends.memcached.PyMemcacheCache` as `pymemcached:` protocol
+* rename `memcached+pylibmccache` protocol to `pylibmccache`
 
 ## 1.3.0